spawn: pass through enemy and pickup depth

+++ src/game/data.rs
@@ -1,2 +1,2 @@
-pub const STARTING_LIVES: usize = 3;
+pub const STARTING_LIVES: u8 = 3;
 pub const DEATH_SCORE_PENALTY: i32 = 150;
